xc refactor : split realization of functionals into 5 separate files
