+++ source/module_xc/CMakeLists.txt
@@ -3,6 +3,11 @@ list(APPEND objects
     potential_libxc_meta.cpp
     xc_functional.cpp
     xc_gga_pw.cpp
+    xc_funct_exch_lda.cpp
+    xc_funct_corr_lda.cpp
+    xc_funct_exch_gga.cpp
+    xc_funct_corr_gga.cpp
+    xc_funct_hcth.cpp
     )
 
 add_library(
